requirements.txt
openai
langchain
llama-index
crewai
chromadb
pymysql
sqlalchemy
python-dotenv
fastapi
uvicorn
streamlit
sqlparse
requests
